Add Tracer trait
